Rebuilds the port on top of the Bazel 9 / rules_lint / spdlog upgrade
that already landed on master, keeping master's choices where the two
overlapped:

- src/common/logging.h stays master's spdlog-backed LOG/CHECK layer; every
  module unit now includes it instead of glog, and //src/common:logging
  replaces the glog dep in BUILD files. The tbox logging test is dropped
  with the tbox logging API.
- master's winsock_init library, WebSocket upgrade handler and the
  self-contained websocket_client_test (in-process proxygen server) are
  carried over: handler_factory.cc routes /api/v1/folder/load to
  WebSocketUpgradeHandler and the test imports the two modules.
- client.cc keeps master's bind-address normalisation and both mains use
  logging::Initialize/CommandLine.
- Repositories nothing referenced are removed from MODULE.bazel with
  their overlays: vlmcsd, sqlite, unordered_dense, blake3, crc32c,
  rapidjson, multipart-parser, simdjson, smhasher, libev and the unused
  bzip2 alias (org_bzip_bzip2 stays for boost and folly).
- clang-tidy 22 findings in the inherited code are fixed: scoped_lock,
  noexcept log destructors, uint8_t stream output, an exception-safe
  test main and a duplicated declaration the merge produced.

Verified on Windows (MSVC 14.52): bazel build --config=lint and
bazel test (8/8) pass.

+++ src/server/http_handler/websocket_handler.h
@@ -122,7 +122,8 @@ class WebSocketHandler {
 
     // Validate opcode
     if (header.opcode > 0xA) {
-      LOG(ERROR) << "Invalid opcode: " << header.opcode;
+      LOG(ERROR) << "Invalid opcode: "
+                 << static_cast<unsigned int>(header.opcode);
       return false;
     }
 
@@ -267,7 +268,8 @@ class WebSocketHandler {
           return true;
 
         default:
-          LOG(ERROR) << "Unknown control frame opcode: " << header.opcode;
+          LOG(ERROR) << "Unknown control frame opcode: "
+                     << static_cast<unsigned int>(header.opcode);
           return false;
       }
     }
